Restructure the project
1. All components were put on the same level in the ./src/components dir
2. All CSS were put in the dir of the components
3. All script use to insert the react components were put in the
   ./src/scripts dir.
4. Other changes
