+++ src/components/main-pane/main-pane.css
