+++ src/components/navigation-pane/navigation-pane.css
