+++ src/components/tab-bar-page/index.ts
@@ -1,0 +1,1 @@
+export * from './tab-bar-page';
