+++ src/components/tab-bar-page/tab-bar-page.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import * as style from '../../styles/index/index.css';
-import { MainPane } from './main-pane';
-import { NavigationPane } from './navigation-pane';
+import * as style from './tab-bar-page.css';
+import { MainPane } from '../main-pane';
+import { NavigationPane } from '../navigation-pane';
 import { initializeIcons } from '@uifabric/icons';
 initializeIcons();
 
-export function Index() {
+export function TabBarPage() {
   return (
     <div className={style.index}>
       <NavigationPane/>
