+++ src/components/tab-group/index.ts
@@ -1,0 +1,1 @@
+export * from './tab-group';
