+++ src/components/tab-item/index.ts
@@ -1,0 +1,1 @@
+export * from './tab-item';
