+++ src/components/tab-item/tab-item.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import * as style from '../../styles/index/tab.css';
+import * as style from './tab-item.css';
 import * as storage from '../../storage';
 
-export function Tab({ tab }: { tab: storage.Tab }) {
+export function TabItem({ tab }: { tab: storage.Tab }) {
   return (
     <div className={style.tab}>
       <div className={style.title}>{tab.name}</div>
