+++ src/content-scripts/tab-bar/index.ts
@@ -1,0 +1,3 @@
+import { insertTabBar } from './tab-bar';
+
+insertTabBar();
