+++ src/content-scripts/tab-bar/tab-bar.css
