+++ src/content-scripts/tab-bar/tab-bar.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import * as ReactDom from 'react-dom';
-import * as style from './index.css';
+import * as style from './tab-bar.css';
 import getBrowserTabId from '../../utils/getBrowserTabId';
 import { Storage, LocalStorage, TabGroup } from '../../storage';
 import { initializeIcons } from '@uifabric/icons';
 import { TabBar } from '../../components/tab-bar/tab-bar';
 
-(async () => {
-  
+export async function insertTabBar() {
+
   initializeIcons();
 
   const storage = new Storage(new LocalStorage());
@@ -18,5 +18,5 @@ import { TabBar } from '../../components/tab-bar/tab-bar';
   root.classList.add(style.reset);
   root.classList.add(style.tabBarContainer);
   document.body.appendChild(root);
-  ReactDom.render(<TabBar tabGroup={tabGroup}/>, root);
-})();
+  ReactDom.render(<TabBar tabGroup={tabGroup} />, root);
+}
