+++ src/scripts/tab-bar-page.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDom from 'react-dom';
-import { Index } from './components/index/index';
+import { TabBarPage } from '../components/tab-bar-page/tab-bar-page';
 
 const root = document.createElement('div');
 document.body.prepend(root);
-ReactDom.render(<Index/>, root);
+ReactDom.render(<TabBarPage/>, root);
